requirements.txt
requests==2.32.3
tqdm==4.66.5
python-dotenv==1.0.1
